requirements.txt
cryptography>=42.0.0
feedparser==6.0.11
fastapi==0.115.8
jinja2==3.1.5
openai==1.68.2
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
uvicorn==0.34.0
youtube-transcript-api>=1.1.0
yt-dlp>=2025.10.14
